Add fullscreen button
The icon doesn't change because the component is never informed of the
change, but this is trivial. I just wanted to get this feature out
because it's something I've been desiring for a while now.

On Android, fullscreen doesn't mean "fullscreen". It just takes up 100%
of the app's viewport, rather than fullscreening like a video player
would. I'd like to work out what to do to achieve true fullscreen but in
the interest of time, here's a quick fullscreen solution.